requirements.txt
SQLAlchemy==2.0.19
python-dotenv==1.0.0
Flask==2.3.2
mysql-connector-python==8.1.0
